+++ ft_strncmp.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int ft_strncmp(const char *s1, const char *s2, int n)
+int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    int uu;
+    unsigned int uu;
 
     uu = 0;
     while (s1[uu] || s2[uu])
